+++ packages/documentation/src/stories/components/accordion/accordion.snapshot.stories.ts
@@ -1,5 +1,5 @@
 import type { StoryContext, StoryObj } from '@storybook/web-components';
-import meta, { Default } from './accordion.stories';
+import meta from './accordion.stories';
 import { html } from 'lit';
 
 const { id, ...metaWithoutId } = meta;
@@ -9,16 +9,19 @@ export default {
   title: 'Snapshots',
 };
 
-type Story = StoryObj<HTMLPostAccordionElement>;
+type Story = StoryObj<HTMLPostAccordionElement & HTMLPostCollapsibleElementEventMap>;
 
 export const Accordion: Story = {
-  render: (_args: HTMLPostAccordionElement, context: StoryContext<HTMLPostAccordionElement>) => {
+  render: (
+    _args: HTMLPostAccordionElement,
+    context: StoryContext<HTMLPostAccordionElement & HTMLPostCollapsibleElementEventMap>,
+  ) => {
     return html`
       <div>
         ${['bg-white', 'bg-light', 'bg-dark'].map(
           bg => html`
             <div class="${bg} d-flex flex-column gap-regular p-regular mt-regular">
-              ${Default.render?.({ ...context.args }, context)}
+              ${meta.render?.({ ...context.args }, context)}
             </div>
           `,
         )}